7/31 Meeting Demonstration
(+) Simultaniously, when clicking a state, the districts of that state are loaded and the map is zoomed and adjusted to look at that state.
(+) When zooming back out after in district view, state view is automatically set. But zooming in from state view does not bring you to district view, that requires a state to be selected.
(+) Pop-ups to again allow for information about a district to be displayed
(+) Changing the colors of individual districts